Move sponsorship detail view to users app
