+++ sponsors/urls.py
@@ -10,9 +10,4 @@ urlpatterns = [
     path('application/', views.SelectSponsorshipApplicationBenefitsView.as_view(),
         name="select_sponsorship_application_benefits",
     ),
-    path(
-        "application/<int:pk>/detail/",
-        views.SponsorshipDetailView.as_view(),
-        name="sponsorship_application_detail",
-    ),
 ]
